Created a file named 'script.js'.
Created a file named 'script.js' to allow users enter their unique identifier, and then to select the mood that they feel from a list (i.e. happy, sad, angry, tired, and scared) via the form that it provides (using AJAX). Additionally, it provides users with feedback (through text) that their data was successfully submitted to the location in which it is stored (i.e. a SQLite database).