Updated 'script.js'.
Updated the file named 'script.js' to check whether or not a unique identifier (that is associated with a particular user) is associated with an existing mood entry for the present day. It now has a five-hundred millisecond delay to prevent excessive API calls. The display system for messages to users now has success, error, and information variants.