Created 'weekly_recap.js'.
Created a file named 'weekly_recap.js'. It determines the dynamic behaviour of the file named 'weekly_recap.html'. Specifically, it attempts to load the data of the unique user identifier that has been entered into its respective page. It displays a spinner while it is doing that. It calls the function named 'updateStats()' to populate particular statistical values (i.e. the days of the week where the respective unique user identifier had the best and worst moods, the mood that they most frequently experienced that week, as well as the one-three most frequently experiences reasons for that mood) when it is successful, and an error message when it fails.