Created 'recap.js'.
Created a file named 'recap.js' to allow users to determine the analyses of data that they are presented with (i.e. either a weekly analysis of data, or a monthly analysis of data). 

It results in the weekly analysis of that data presenting its user with the most frequently occurring mood for a user over a respective week, as well as the most frequently occurring activities done by that user over that week. 

It results in the monthly analysis of that data presenting its user with the most frequently occurring day over a respective month that a user has felt every mood (i.e. happy, sad, angry, tired, and scared), as well as the most frequently occurring reasons for that (i.e. from 'Went outside.', 'Took care of their health.', 'Played video games.', 'School/work/study/productivity.', 'Had good food or drinks.', 'Had bad food or drinks.', 'Travelled.', 'Socialised.', 'Listened to music.', 'Watched a film or television show.', 'Read.', 'Did art.', 'Rested.', 'Did errands.'.