Created 'calendar.html'.
Created a file named 'calendar.html' that provides the weekly and monthly views of the calendars that display users' respective mood data. It provides colour-coded indicators for those moods on those calendars.. It provides tags for the respective reasons for users' moods on those calendars. It provides access to users' relationships' calendars.