Updated 'main.py'.
Updated the file named 'main.py' to return all of the most frequently (i.e. occurring three or more times within a timeframe) activities that are the reasons for a respective users' moods. 'None' is presented if that is not applicable for that respective user. Additionally, it now tracks the day of the week of a respective date. Additionally, 'username' is used in place of all instances of 'name'.