Updated 'main.py'.
Updated the file named 'main.py' by adding a new route ('/delete_account') to it, with POST method. That route is protected with the '@login_required' decorator. It allows a cascading deletion of all of a user's moods, all of that user's friendships (both directions), all of that use's friend requests (both sent and received), as well as that account itself. When it is successful, it will provide a message and then redirect to the login page of the application.