Updated 'changepassword.html'.
Updated the file named 'changepassword.html' to add a section to it that allows its users to delete their respective accounts. It has a warning message. It has a button which triggers a model. It has the ability for its following actions be cancelled. It is a form that submits to the 'delete_account' route.