Updated 'home.html'.
Updated the file named 'home.html' by adding a search bar in the 'Friends' sidebar. Additionally, JavaScript functions were added for searching users, displaying the results of that search (in the same format as 'Friends' sidebar), sending friend requests, as well as for toggling between search results and the 'Friends' sidebar.