Created 'testmodels.py'.
Creation of a file named 'testmodels.py', to contains tests for the 'User' and 'Mood' models. It tests whether or not a new user can be created. and then stored in the database. It tests whether or not the password hashing and verification functionalities work correctly. It tests whether or not a mood can be created, and then associated with a user.